refactor(cache): move cache functions from `router` crate to `storage_impl` crate
