+++ crates/router/src/db/merchant_connector_account.rs
@@ -203,7 +203,7 @@ impl MerchantConnectorAccountInterface for Store {
 
         #[cfg(feature = "accounts_cache")]
         {
-            super::cache::get_or_populate_in_memory(
+            cache::get_or_populate_in_memory(
                 self,
                 &format!("{}_{}", merchant_id, connector_label),
                 find_call,
@@ -248,7 +248,7 @@ impl MerchantConnectorAccountInterface for Store {
 
         #[cfg(feature = "accounts_cache")]
         {
-            super::cache::get_or_populate_in_memory(
+            cache::get_or_populate_in_memory(
                 self,
                 &format!("{}_{}", profile_id, connector_name),
                 find_call,
@@ -322,7 +322,7 @@ impl MerchantConnectorAccountInterface for Store {
 
         #[cfg(feature = "accounts_cache")]
         {
-            super::cache::get_or_populate_in_memory(
+            cache::get_or_populate_in_memory(
                 self,
                 &format!("{}_{}", merchant_id, merchant_connector_id),
                 find_call,
@@ -418,7 +418,7 @@ impl MerchantConnectorAccountInterface for Store {
         #[cfg(feature = "accounts_cache")]
         {
             // Redact both the caches as any one or both might be used because of backwards compatibility
-            super::cache::publish_and_redact_multiple(
+            cache::publish_and_redact_multiple(
                 self,
                 [
                     cache::CacheKind::Accounts(
@@ -481,7 +481,7 @@ impl MerchantConnectorAccountInterface for Store {
                 "profile_id".to_string(),
             ))?;
 
-            super::cache::publish_and_redact_multiple(
+            cache::publish_and_redact_multiple(
                 self,
                 [
                     cache::CacheKind::Accounts(
@@ -766,6 +766,7 @@ impl MerchantConnectorAccountInterface for MockDb {
     }
 }
 
+#[cfg(feature = "accounts_cache")]
 #[cfg(test)]
 mod merchant_connector_account_cache_tests {
     use api_models::enums::CountryAlpha2;
@@ -774,7 +775,7 @@ mod merchant_connector_account_cache_tests {
     use error_stack::ResultExt;
     use masking::PeekInterface;
     use storage_impl::redis::{
-        cache::{CacheKind, ACCOUNTS_CACHE},
+        cache::{self, CacheKind, ACCOUNTS_CACHE},
         kv_store::RedisConnInterface,
         pub_sub::PubSubInterface,
     };
@@ -783,7 +784,7 @@ mod merchant_connector_account_cache_tests {
     use crate::{
         core::errors,
         db::{
-            cache, merchant_connector_account::MerchantConnectorAccountInterface,
+            merchant_connector_account::MerchantConnectorAccountInterface,
             merchant_key_store::MerchantKeyStoreInterface, MasterKeyInterface, MockDb,
         },
         services,
